refactor: wrap next PageLoader via `patchNextClientPageLoader` which init global `mf_loader` variable and import `include-dafaults`.
This refactor helps to simplify app code
- remove `.real` async pages
- now you just need add NextFederationPlugin to next.config.js, nothing else
