+++ demo/3001-shop/next.config.js
@@ -3,6 +3,8 @@ const NextFederationPlugin = require('@module-federation/nextjs-mf');
 module.exports = {
   webpack(config, options) {
     if (!options.isServer) {
+      config.module.rules.push(...NextFederationPlugin.loaders);
+
       config.plugins.push(
         new NextFederationPlugin({
           name: 'shop',
