+++ demo/3002-checkout/next.config.js
@@ -2,10 +2,9 @@ const NextFederationPlugin = require('@module-federation/nextjs-mf');
 
 module.exports = {
   webpack(config, options) {
-    const { webpack } = options;
-    Object.assign(config.experiments, { topLevelAwait: true });
     if (!options.isServer) {
-      //config.cache=false
+      config.module.rules.push(...NextFederationPlugin.loaders);
+
       config.plugins.push(
         new NextFederationPlugin({
           name: 'checkout',
