+++ src/include-defaults.js
@@ -13,6 +13,4 @@ if (process.env.NODE_ENV === 'development') {
   require('react/jsx-dev-runtime');
 }
 
-require('./mf-router');
-
 module.exports = {};
